Renamed cfengine.py to cfengine_module_library.py
