+++ examples/git-using-lib/git_using_lib.py
@@ -1,5 +1,5 @@
 import os
-from cfengine import PromiseModule, ValidationError, Result
+from cfengine_module_library import PromiseModule, ValidationError, Result
 
 
 class GitPromiseTypeModule(PromiseModule):
