+++ examples/gpg/gpg.py
@@ -44,7 +44,8 @@ bundle agent main
 import json
 from subprocess import Popen, PIPE
 import sys
-from cfengine import PromiseModule, ValidationError, Result
+from cfengine_module_library import PromiseModule, ValidationError, Result
+
 
 class GpgKeysPromiseTypeModule(PromiseModule):
     def __init__(self):
