+++ libraries/python/cfengine_module_library.py
