+++ promise-types/http/http_promise_type.py
@@ -8,7 +8,7 @@ import ssl
 import json
 from contextlib import contextmanager
 
-from cfengine import PromiseModule, ValidationError, Result
+from cfengine_module_library import PromiseModule, ValidationError, Result
 
 
 _SUPPORTED_METHODS = {"GET", "POST", "PUT", "DELETE", "PATCH"}
